+++ src/client/main.jsx
@@ -5,6 +5,8 @@ import { createStore, applyMiddleware } from 'redux';
 import { Provider } from 'react-redux';
 import thunk from 'redux-thunk';
 import * as Immutable from 'immutable';
+import { createBrowserHistory } from 'history';
+import { connectRouter, routerMiddleware } from 'connected-react-router/immutable';
 
 import { App } from './app';
 import { rootReducer } from './app/reducers';
@@ -15,16 +17,21 @@ import './main.scss';
 // favicon
 import './favicon.ico';
 
+const history = createBrowserHistory();
 const initialState = Immutable.Map();
+
 const store = createStore(
-  rootReducer,
+  connectRouter(history)(rootReducer),
   initialState,
-  applyMiddleware(thunk)
+  applyMiddleware(
+    routerMiddleware(history),
+    thunk
+  )
 );
 
 render(
   <Provider store={store}>
-    <App />
+    <App history={history} />
   </Provider>,
   document.getElementById('app')
 );